pyproject.toml
[tool.poetry]
name = "onacol"
version = "0.1.0"
description = "Oh No! Another Configuration Library"
authors = ["Josef Nevrly"]
license = "MIT license"

[tool.poetry.dependencies]
python = "^3.6"
Cerberus = "^1.3.4"
"ruamel.yaml" = "^0.17.10"
cascadict = "^0.8.4"
attrs = "^21.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
